+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - [UPGRADED] Minimum Swift language version to 4.2.
 - [UPGRADED] Swift tools version and package manifest to version 5.0.
 - [BREAKING CHANGE] Projects using older Swift tool and language versions will not work with versions newer than 0.8.0.
+- [FIXED] Fixed scope of `Operation.Error` to be exposed as public.
 
 # 0.8.0 (2018-03-20)